requirements.txt
ipykernel
ipywidgets

pymilvus
numpy
mlserver

transformers
torch

langchain
langchain_community
sentence_transformers

fastapi
uvicorn
uvicorn[standard]

gradio

rank_bm25
psycopg2-binary

openai